refactor: move UserDataStore to UserService
